format squad_train_model.py back into original form
